pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Balaji Seetharaman"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.3"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.5"
whitenoise = {extras = ["brotli"], version = "^5.1.0"}
gunicorn = "^20.0.4"
youtube-transcript-api = "^0.6.0"
mindsdb_sdk = "^1.0.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
